Added authentication logic for users and non-users

+++ app/controllers/products_controller.rb
@@ -4,6 +4,9 @@ class ProductsController < ApplicationController
   end
 
   def index
+
+    # calculate_cart_count
+
     @products = Product.all
 
     sort_attribute = params[:input_sort]
@@ -34,47 +37,66 @@ class ProductsController < ApplicationController
   end
 
   def new
+    unless current_user && current_user.admin
+      redirect_to "/"
+    end
   end
 
   def create
-    product = Product.create(
-      name: params[:name],
-      price: params[:price],
-      description: params[:description],
-      deliciousness: params[:deliciousness],
-      stocked: params[:stocked],
-      quantity: params[:quantity],
-      image: params[:image]
-      )
-    flash[:success] = "Product created. Way to go guy!"
-    redirect_to '/products'
+    if current_user && current_user.admin
+      product = Product.create(
+        name: params[:name],
+        price: params[:price],
+        description: params[:description],
+        deliciousness: params[:deliciousness],
+        stocked: params[:stocked],
+        quantity: params[:quantity],
+        image: params[:image]
+        )
+      flash[:success] = "Product created. Way to go guy!"
+      redirect_to '/products'
+    else
+      redirect_to "/"
+    end
   end
 
   def edit
-    product_id = params[:id]
-    @product = Product.find_by(id: params[:id])
+    if current_user && current_user.admin
+      product_id = params[:id]
+      @product = Product.find_by(id: params[:id])
+    else
+      redirect_to "/"
+    end
   end
 
   def update
-    product = Product.find_by(id: params[:id])
-    product.update(
-      name: params[:name],
-      price: params[:price],
-      description: params[:description],
-      deliciousness: params[:deliciousness],
-      stocked: params[:stocked],
-      quantity: params[:quantity],
-      image: params[:image]
-      )
-    flash[:info] = "Product updated. Way to go friend!"
-    redirect_to "/products/#{product.id}"
+    if current_user && current_user.admin
+      product = Product.find_by(id: params[:id])
+      product.update(
+        name: params[:name],
+        price: params[:price],
+        description: params[:description],
+        deliciousness: params[:deliciousness],
+        stocked: params[:stocked],
+        quantity: params[:quantity],
+        image: params[:image]
+        )
+      flash[:info] = "Product updated. Way to go friend!"
+      redirect_to "/products/#{product.id}"
+    else
+      redirect_to "/"
+    end
   end
 
   def destroy
-    product = Product.find_by(id: params[:id])
-    product.destroy!
-    flash[:danger] = "Product eliminated. Way to go buddy!"
-    redirect_to '/products'
+    if current_user && current_user.admin
+      product = Product.find_by(id: params[:id])
+      product.destroy!
+      flash[:danger] = "Product eliminated. Way to go buddy!"
+      redirect_to '/products'
+    else
+      redirect_to "/"
+    end
   end
 
   def checkout
